chore(srpc): update package configuration and enhance client/server bridges
- Bump version to 0.1.1 and set module type in package.json.
- Modify tsconfig to include test files.
- Update entry points in tsup.config for client and server implementations.
- Introduce Zod-enabled client and server bridges for enhanced type safety.
- Add setup for browser environment in tests and adjust test imports.
- Remove deprecated type helpers file to streamline the codebase.
