+++ packages/srpc/src/client.ts
@@ -1,94 +1,27 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-// WebSocket-based RPC client implementation for type-safe remote procedure calls
-// Copyright (C) 2025 Goetze, Scharpff & Toews GbR
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as
-// published by the Free Software Foundation, either version 3 of the
-// License, or (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
 import { WebSocketRpcBridge, type WebSocketBridgeOptions } from './core';
-
-// Use the appropriate WebSocket implementation based on the environment
-const WebSocketImpl =
-  typeof window !== 'undefined' ? window.WebSocket : require('ws').WebSocket;
+import { ZodTypedBridge } from './zod-bridge';
+import type { ZodBridgeContract } from './zod-contract';
 
 /**
- * Client implementation of the WebSocket RPC Bridge
+ * Extended WebSocketRpcBridge for client-side use
  */
-export class WebSocketRpcClient extends WebSocketRpcBridge {
+class ClientBridge extends WebSocketRpcBridge {
   private url: string;
   private reconnectTimer: NodeJS.Timeout | null = null;
 
-  /**
-   * Creates a new WebSocketRpcClient
-   * @param url WebSocket server URL
-   * @param options Connection options
-   */
   constructor(url: string, options?: WebSocketBridgeOptions) {
     super(options);
     this.url = url;
   }
 
-  /**
-   * Connect to the WebSocket server
-   * @returns Promise that resolves when connected
-   */
-  public connect(): Promise<void> {
-    return new Promise((resolve, reject) => {
-      try {
-        const ws = new WebSocketImpl(this.url);
-
-        ws.onopen = () => {
-          this.ws = ws;
-          this.reconnectAttempts = 0;
-          this.isIntentionalClose = false;
-          this.setupWebSocketHandlers(ws);
-          resolve();
-        };
-
-        ws.onerror = (_event: Event) => {
-          reject(new Error('Failed to connect to WebSocket server'));
-        };
-      } catch (error) {
-        reject(error);
-      }
-    });
-  }
-
-  /**
-   * Method to call a remote procedure on the server
-   * @param method Method name
-   * @param payload Request payload
-   * @param onUpdate Optional callback for updates
-   * @returns Promise resolving with the response
-   */
   public call<TRequest, TResponse, TUpdate>(
     method: string,
     payload: TRequest,
     onUpdate?: (update: TUpdate) => void,
   ): Promise<TResponse> {
-    if (!this.ws) {
-      throw new Error('WebSocket is not connected. Call connect() first.');
-    }
-    return this.callMethod<TRequest, TResponse, TUpdate>(
-      method,
-      payload,
-      onUpdate,
-    );
+    return this.callMethod(method, payload, onUpdate);
   }
 
-  /**
-   * Reconnect to the WebSocket server
-   */
   protected reconnect(): void {
     if (this.reconnectTimer) {
       clearTimeout(this.reconnectTimer);
@@ -98,8 +31,59 @@ export class WebSocketRpcClient extends WebSocketRpcBridge {
       try {
         await this.connect();
       } catch (error) {
-        this.handleDisconnect();
+        this.reconnect(); // Try again
       }
     }, this.options.reconnectDelay);
   }
+
+  public connect(): Promise<void> {
+    return new Promise((resolve, reject) => {
+      try {
+        const ws = new window.WebSocket(this.url);
+
+        ws.onopen = () => {
+          this.ws = ws;
+          this.setupWebSocketHandlers(ws);
+          resolve();
+        };
+
+        ws.onerror = () => {
+          reject(new Error('Failed to connect to WebSocket server'));
+        };
+      } catch (error) {
+        reject(error);
+      }
+    });
+  }
+}
+
+/**
+ * Zod-enabled client bridge
+ */
+export class ZodClient<C extends ZodBridgeContract> extends ZodTypedBridge<
+  NonNullable<C['client']>,
+  NonNullable<C['server']>,
+  ClientBridge
+> {
+  constructor(url: string, contract: C, options?: WebSocketBridgeOptions) {
+    super(new ClientBridge(url, options), {
+      serves: contract.client || ({} as NonNullable<C['client']>),
+      consumes: contract.server || ({} as NonNullable<C['server']>),
+    });
+  }
+
+  public connect(): Promise<void> {
+    return (this.bridge as ClientBridge).connect();
+  }
+}
+
+/**
+ * Helper function to create a Zod-enabled client bridge
+ */
+export function createSRPCClientBridge<C extends ZodBridgeContract>(
+  url: string,
+  contract: C,
+  options?: WebSocketBridgeOptions,
+): ZodClient<C> {
+  return new ZodClient(url, contract, options);
 }
