+++ packages/srpc/src/index.ts
@@ -30,31 +30,16 @@ export {
   type WebSocketMessage,
   type WebSocketBridgeOptions,
   type PendingRequest,
-} from './src/core';
+} from './core';
 
-// Export the server
-export { WebSocketRpcServer } from './src/server';
-
-// Export the client
-export { WebSocketRpcClient } from './src/client';
-
-// Export type helpers
-export {
-  type EndpointMethodMap,
-  type RpcMethodContract,
-  type MethodImplementations,
-  type MethodCalls,
-  TypedBridge,
-  TypedServer,
-  TypedClient,
-  type CreateBridgeContract,
-} from './src/type-helpers';
-
-export { createBridgeContract } from './src/zod-contract';
+export { createBridgeContract } from './zod-contract';
 
 export {
   ZodServer,
-  ZodClient,
   createSRPCServerBridge,
+} from './server';
+
+export {
+  ZodClient,
   createSRPCClientBridge,
-} from './src/zod-bridge';
+} from './client';
