+++ packages/srpc/src/zod-bridge.ts
@@ -15,16 +15,9 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-import { WebSocketRpcBridge, type WebSocketBridgeOptions } from './core';
-import type { Server } from 'node:http';
-import { WebSocketServer, type WebSocket as NodeWebSocket } from 'ws';
-
-// Use the appropriate WebSocket implementation based on the environment
-const WebSocketImpl =
-  typeof window !== 'undefined' ? window.WebSocket : require('ws').WebSocket;
+import type { WebSocketRpcBridge } from './core';
 
 import type {
-  ZodBridgeContract,
   ZodEndpointMethodMap,
   ZodMethodCalls,
   ZodMethodImplementations,
@@ -173,161 +166,4 @@ export class ZodTypedBridge<
   public async close(): Promise<void> {
     await this.bridge.close();
   }
-}
-
-/**
- * Extended WebSocketRpcBridge for server-side use
- */
-class ServerBridge extends WebSocketRpcBridge {
-  private wss: WebSocketServer;
-
-  constructor(server: Server) {
-    super();
-    this.wss = new WebSocketServer({ server });
-
-    this.wss.on('connection', (ws: NodeWebSocket) => {
-      if (this.ws) {
-        console.warn(
-          'New WebSocket connection attempted while one is already active. Closing existing connection first.',
-        );
-        const oldWs = this.ws;
-        this.ws = null;
-        oldWs.close();
-      }
-
-      this.ws = ws;
-      this.setupWebSocketHandlers(ws);
-
-      ws.on('close', () => {
-        if (this.ws === ws) {
-          this.ws = null;
-        }
-      });
-    });
-  }
-
-  public call<TRequest, TResponse, TUpdate>(
-    method: string,
-    payload: TRequest,
-    onUpdate?: (update: TUpdate) => void,
-  ): Promise<TResponse> {
-    return this.callMethod(method, payload, onUpdate);
-  }
-
-  // Server does not need to reconnect
-  protected reconnect(): void {}
-
-  public close(): Promise<void> {
-    this.wss.close();
-    return super.close();
-  }
-}
-
-/**
- * Extended WebSocketRpcBridge for client-side use
- */
-class ClientBridge extends WebSocketRpcBridge {
-  private url: string;
-  private reconnectTimer: NodeJS.Timeout | null = null;
-
-  constructor(url: string, options?: WebSocketBridgeOptions) {
-    super(options);
-    this.url = url;
-  }
-
-  public call<TRequest, TResponse, TUpdate>(
-    method: string,
-    payload: TRequest,
-    onUpdate?: (update: TUpdate) => void,
-  ): Promise<TResponse> {
-    return this.callMethod(method, payload, onUpdate);
-  }
-
-  protected reconnect(): void {
-    if (this.reconnectTimer) {
-      clearTimeout(this.reconnectTimer);
-    }
-
-    this.reconnectTimer = setTimeout(async () => {
-      try {
-        await this.connect();
-      } catch (error) {
-        this.reconnect(); // Try again
-      }
-    }, this.options.reconnectDelay);
-  }
-
-  public connect(): Promise<void> {
-    return new Promise((resolve, reject) => {
-      try {
-        const ws = new WebSocketImpl(this.url);
-
-        ws.onopen = () => {
-          this.ws = ws;
-          this.setupWebSocketHandlers(ws);
-          resolve();
-        };
-
-        ws.onerror = () => {
-          reject(new Error('Failed to connect to WebSocket server'));
-        };
-      } catch (error) {
-        reject(error);
-      }
-    });
-  }
-}
-
-/**
- * Zod-enabled server bridge
- */
-export class ZodServer<C extends ZodBridgeContract> extends ZodTypedBridge<
-  NonNullable<C['server']>,
-  NonNullable<C['client']>,
-  ServerBridge
-> {
-  constructor(server: Server, contract: C) {
-    super(new ServerBridge(server), {
-      serves: contract.server || ({} as NonNullable<C['server']>),
-      consumes: contract.client || ({} as NonNullable<C['client']>),
-    });
-  }
-}
-
-/**
- * Zod-enabled client bridge
- */
-export class ZodClient<C extends ZodBridgeContract> extends ZodTypedBridge<
-  NonNullable<C['client']>,
-  NonNullable<C['server']>,
-  ClientBridge
-> {
-  constructor(url: string, contract: C, options?: WebSocketBridgeOptions) {
-    super(new ClientBridge(url, options), {
-      serves: contract.client || ({} as NonNullable<C['client']>),
-      consumes: contract.server || ({} as NonNullable<C['server']>),
-    });
-  }
-
-  public connect(): Promise<void> {
-    return (this.bridge as ClientBridge).connect();
-  }
-}
-
-/**
- * Helper functions to create Zod-enabled bridges
- */
-export function createSRPCServerBridge<C extends ZodBridgeContract>(
-  server: Server,
-  contract: C,
-): ZodServer<C> {
-  return new ZodServer(server, contract);
-}
-
-export function createSRPCClientBridge<C extends ZodBridgeContract>(
-  url: string,
-  contract: C,
-  options?: WebSocketBridgeOptions,
-): ZodClient<C> {
-  return new ZodClient(url, contract, options);
 }
