+++ packages/srpc/tsup.config.ts
@@ -18,7 +18,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['./index.ts'],
+  entry: ['./src/index.ts', './src/client.ts', './src/server.ts'],
   format: ['cjs', 'esm'],
   dts: true,
   clean: true,
