+++ packages/srpc/vitest.config.ts
@@ -21,5 +21,6 @@ export default defineConfig({
   test: {
     environment: 'node',
     include: ['tests/**/*.test.ts'],
+    setupFiles: ['tests/setup-browser-env.ts'],
   },
 });
